neXt education app Source Code
This is the source code for the neXt education app that I independently programmed. The computer program is in the computer language JavaScript and was built on the Code.org App Lab platform. The data and design elements that the computer program references can be found in the full code with the "</> Show Text" version at https://studio.code.org/projects/applab/aW44MpXBKlXIjr8gfePdq35Hkyd1Ht6_h3-EsuTLKYM/view.